Did some work on the price for executed orders that were partially executed first

+++ app/src/main/java/org/axonframework/samples/trader/app/query/transaction/TransactionEntry.java
@@ -38,7 +38,7 @@ public class TransactionEntry {
         return amountOfExecutedItems;
     }
 
-    public void setAmountOfExecutedItems(int amountOfExecutedItems) {
+    public void setAmountOfExecutedItems(long amountOfExecutedItems) {
         this.amountOfExecutedItems = amountOfExecutedItems;
     }
